🔥 Remove unused code, remove notification and unused permissions in chrome extension manifest file
