+++ src/lib/Pages/Settings.svelte
@@ -2,7 +2,7 @@
 	import Icon from '@iconify/svelte';
 	import ToggleSwitch from '../components/ToggleSwitch.svelte';
 	import { AppPageItem } from '$lib/components/App';
-	import { showNotification, userProfile } from '$lib/stores/data';
+	import { userProfile } from '$lib/stores/data';
 	import type { Relay } from '$lib/types';
 	import { profileController } from '$lib/controllers/profile.controller';
 
@@ -17,10 +17,6 @@
 	const removeRelay = (relay: Relay) => {
 		profileController.removeRelayFromProfile(relay);
 	};
-
-	function handleToggleNotification(event: CustomEvent) {
-		showNotification.set(event.detail);
-	}
 </script>
 
 <AppPageItem name="settings">
@@ -76,10 +72,6 @@
 			class="text-gray-800 dark:text-gray-400 text-opacity-70 text-xs font-semibold leading-4 tracking-[2.4000000000000004px]"
 		>
 			General settings
-		</div>
-		<div class="justify-between items-stretch flex gap-5 mt-2">
-			<div class="text-black dark:text-white text-base leading-5 my-auto">Enable Notification</div>
-			<ToggleSwitch bind:isToggled={$showNotification} on:change={handleToggleNotification} />
 		</div>
 	</div>
