+++ src/lib/controllers/index.ts
@@ -1,0 +1,1 @@
+export * from './browser.controller';
